Close most of the remaining sample-level gap to the reference
Three more places where the port's arithmetic drifted from
`vhsdecode.hifi`, found by dumping per-stage arrays from an instrumented
run on a real capture and comparing stage by stage. Together they take
the residual on a 12s 8fsc decode from 2.14% / 0.25% of L/R samples
differing to 0.63% / 0.007%.

`np.mean` again, in two more places. `_fill`'s three DC measurements and
`smooth`'s moving average both go through it, so both needed the same
reassociated f32 summation `cancelDC_trim` already got. Dropout
compensation on a real block went from 54591 of 96000 samples differing
to none at all.

`_fill`'s ramp and crossfades ran in f32 where the reference runs them in
f64. Python's fade factors are `int / int` and its `denom` is a Python
float, so the raised-cosine ramp, both crossfade products and their sums
are all float64 there, rounding once when they land in the float32 array.
Doing the whole thing in f32 rounds at every step instead.

The DC blocker was never primed. Python runs it over a throwaway copy of
block 0 before the real pass so its three cascaded one-pole stages settle
(`PostProcessor.py:280-284`); the expander priming was ported but this one
was missed, leaving the first block carrying a cold-start transient. That
alone was a third of the remaining L-channel difference.

Also fixes the expander's RMS detector squaring in f64 where numba squares
in f32, and matches scipy's grouping in the head-switch interpolator
(`slope * (x - x_lo) + y_lo` over a float32 `y`, not
`y_lo + (y_hi - y_lo) * t` widened first). Neither is on the default path
or was the dominant term, but both are real.

What is left is one identified cause: head-switch peak detection selects
different peaks. On a probe block Python interpolates around samples 5162,
8770, 12424, 18260 where this port picks 5201, 8807, 12429, 17950 — not a
rounding difference but a different selection, so the fix belongs in the
detector (the cheby2 highpass, `mean_stddev`, or `find_peaks`' distance
handling) rather than in the interpolation. The audible effect is nil (the
surviving differences are 1-2 LSB at 24 bits) but it is the last thing
between this and byte-identical output.

+++ crates/hifi-decode/src/postprocess.rs
@@ -230,7 +230,13 @@ impl Expander {
 
         if self.use_rms {
             for sc in side_chain.iter_mut() {
-                let sc_sq = (*sc as f64) * (*sc as f64);
+                // Squared in `f32`, then widened — not widened and squared
+                // in `f64`. `side_chain` is a float32 array on the Python
+                // side, so `sc * sc` is a float32 multiply there
+                // (`HiFiDecode.py:981`) and rounds before anything else
+                // sees it; doing the multiply in f64 keeps precision the
+                // reference threw away and drifts the envelope from it.
+                let sc_sq = (*sc * *sc) as f64;
                 if sc_sq > env_lin {
                     env_lin = self.atk_coeff * env_lin + one_minus_atk * sc_sq;
                     hold_state = self.hold_samples;
